requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.1
google-generativeai==0.8.3
supabase==2.4.6
pydantic==2.9.2
httpx==0.27.0
